reconstruct active_control and passive_control
