+++ ihealth/RFPassiveTrainAction.cpp
@@ -38,7 +38,7 @@ void OnTimer(HWND hWnd, UINT uMsg, UINT_PTR idEvent, DWORD dwTime)
 			action->SaveMoveingData();
 			if (action->m_medias.size() > 0) {
 				if (action->m_orderplay) {
-					action->m_curmedia = action->PopOrderMeida();
+					action->m_curmedia = action->PopOrderMedia();
 				} else {
 					action->m_curmedia = action->PopAutoMedia();
 				}
@@ -85,7 +85,7 @@ void RFPassiveTrainAction::StartPlay(std::list<MEDIA>& medias, bool orderplay)
 	m_orderplay = orderplay;
 	m_timeplay = 0;
 	if (orderplay) {
-		m_curmedia = PopOrderMeida();
+		m_curmedia = PopOrderMedia();
 	} else {
 		m_curmedia = PopAutoMedia();
 	}
@@ -110,7 +110,7 @@ void RFPassiveTrainAction::PlayNext(bool orderplay)
 
 	m_orderplay = orderplay;
 	if (m_orderplay) {
-		m_curmedia = PopOrderMeida();
+		m_curmedia = PopOrderMedia();
 	} else {
 		m_curmedia = PopAutoMedia();
 	}
@@ -150,7 +150,7 @@ void RFPassiveTrainAction::SetPlayOrder(bool orderplay)
 MEDIA RFPassiveTrainAction::PopAutoMedia()
 {
 	if (m_medias.size() < 2) {
-		return PopOrderMeida();
+		return PopOrderMedia();
 	}
 
 	int nRandomIndex = 0;
@@ -186,7 +186,7 @@ MEDIA RFPassiveTrainAction::PopAutoMedia()
 	return media;
 }
 
-MEDIA RFPassiveTrainAction::PopOrderMeida()
+MEDIA RFPassiveTrainAction::PopOrderMedia()
 {
 	MEDIA media;
 
@@ -206,7 +206,7 @@ MEDIA RFPassiveTrainAction::PopOrderMeida()
 		::Sleep(300U);
 		RFMainWindow::MainWindow->m_robot.startPasvMove(index);
 	} else {
-		PopOrderMeida();
+		PopOrderMedia();
 	}
 
 	return media;
